Update get_ssfr_flight_albedo.m
Cosmetic changes to improve a flight-level albedo plot.  Now is a 3-panel plot incorporating multiple time ranges in panels 1 and 2, and a full-spectral view in panel 3.  Could add flight altitude as a 2nd y-axis for panels 1 and 2.

This change has only been implemented for the part involving the ICT file.  Should propagate to other SSFR formats as well.