GRASP related updates
Changes related to GRASP, especially to figures automatically use a legend for few WLs or a rainbow colorbar for many WLs (>7).  Also, fixed a selection bug in handscreen_skyscan_menu.  Also added ability to select which WLs to use for the retrieval, and also which WLs to use for the polyfit.