Update starsky_scan.m
This version keeps the xfit calls but ultimately uses the original polyfit results for consistency with Kristina's paper.  The polyfit over a limited wavelength range does as well as or better than the aod_basis / xfit stuff over the same limited range.  So xfit stuff may not be necessary unless the WL range is large.  This still to be examined.